Refactor: Made Navbar and User Dropdown components to be re-used on the dashboard page
