--- src/app/home/components/Navbar/index.tsx
@@ -1,25 +1,0 @@
-import Image from "next/image";
-
-import RightOptions from "./Buttons";
-import Link from "next/link";
-
-export default function Navbar() {
-  return (
-    <div className="flex backdrop-blur-3xl w-full z-40 h-18 justify-center border-b-1">
-      <div className="w-full max-w-[100rem] px-5 lg:px-16 xl:px-20 flex justify-between items-center">
-        <div className="flex">
-          <Link href={"/home"}>
-            <Image
-              src={"brand/logo.svg"}
-              width={80}
-              height={80}
-              alt="Logo"
-              className="invert dark:invert-0"
-            />
-          </Link>
-        </div>
-        <RightOptions />
-      </div>
-    </div>
-  );
-}
