+++ src/app/home/layout.tsx
@@ -1,11 +1,11 @@
 import HomepageFooter from "./components/HomepageFooter";
-import Navbar from "./components/Navbar";
+import HomepageNavbar from "./components/HomepageNavbar";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <div className="flex flex-col justify-between min-h-dvh">
       <header className="sticky top-0 flex shrink-0">
-        <Navbar />
+        <HomepageNavbar />
       </header>
       <main className="flex-1 overflow-auto bg-[radial-gradient(var(--secondary),transparent_1px)] [background-size:16px_16px]">
         {children}
