+++ src/components/Navbar/UserDropdown.tsx
@@ -1,0 +1,89 @@
+"use client";
+import {
+  LayoutDashboard,
+  LogOut,
+  Moon,
+  Settings,
+  Sun,
+  SunMoon,
+  User,
+} from "lucide-react";
+import { Button } from "../ui/button";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuGroup,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuRadioGroup,
+  DropdownMenuRadioItem,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "../ui/dropdown-menu";
+import Link from "next/link";
+import { useAuth } from "@/src/hooks/useAuth";
+import { useTheme } from "next-themes";
+import { signOut } from "firebase/auth";
+import { auth } from "@/src/lib/firebase/firebaseClient";
+import { Skeleton } from "../ui/skeleton";
+
+export default function UserDropdown() {
+  const { user, loading } = useAuth();
+  const { setTheme, theme } = useTheme();
+
+  if (loading || !user) {
+    return <Skeleton className="size-8" />;
+  }
+
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button variant={"outline"} size={"icon"}>
+          <User />
+        </Button>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent className="w-56" align="end">
+        <DropdownMenuLabel>
+          <div className="flex gap-1 items-center">
+            <User size={16} /> {user!.email}
+          </div>
+        </DropdownMenuLabel>
+        <DropdownMenuSeparator />
+        <DropdownMenuGroup>
+          <DropdownMenuItem asChild>
+            <Link href={"/dashboard"}>
+              <LayoutDashboard />
+              <span>Dashboard</span>
+            </Link>
+          </DropdownMenuItem>
+          <DropdownMenuItem asChild>
+            <Link href={"/account"}>
+              <Settings />
+              <span>Account Preferences</span>
+            </Link>
+          </DropdownMenuItem>
+        </DropdownMenuGroup>
+        <DropdownMenuSeparator />
+        <DropdownMenuRadioGroup value={theme} onValueChange={setTheme}>
+          <DropdownMenuRadioItem value="light">
+            <Sun />
+            <span>Light</span>
+          </DropdownMenuRadioItem>
+          <DropdownMenuRadioItem value="dark">
+            <Moon />
+            <span>Dark</span>
+          </DropdownMenuRadioItem>
+          <DropdownMenuRadioItem value="system">
+            <SunMoon />
+            <span>System</span>
+          </DropdownMenuRadioItem>
+        </DropdownMenuRadioGroup>
+        <DropdownMenuSeparator />
+        <DropdownMenuItem onClick={() => signOut(auth)}>
+          <LogOut />
+          <span>Logout</span>
+        </DropdownMenuItem>
+      </DropdownMenuContent>
+    </DropdownMenu>
+  );
+}
